% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromo.Crawl.R
\name{Chromo.Crawl}
\alias{Chromo.Crawl}
\title{Chromo.Crawl: (original version) function to crawl over an alignment of chromosomes while attempting to concatenate adjacent genomic windows}
\usage{
Chromo.Crawl(
  string.PathParentDir,
  numeric.WindowSize,
  numeric.StepSize,
  string.PathToFastaFile,
  numeric.NumberOfCores
)
}
\arguments{
\item{string.PathParentDir}{Path to parent dir for all analyses}

\item{numeric.WindowSize}{Size of each chromosomal window}

\item{numeric.StepSize}{Spacing of each chromosomal window}

\item{string.PathToFastaFile}{Path to fasta file}

\item{numeric.NumberOfCores}{Number of cores}
}
\value{
Results Results written to outputfiles
}
\description{
This function writes the results to an output file. Additionally, this function provides a directory where the resulting supergenes are placed.
}
\examples{



#####################################
# Read example chromosome alignment #
#####################################
String.Path_ExampleChromosomeAlignment <-  system.file("extdata", "Example_Chr10.fasta", package="PhyloWGA")

#########################
# Conduct.ChromoPhylome #
#########################
Chromo.Crawl(string.PathParentDir = '~/Desktop/',
                         numeric.WindowSize = 1000,
                         numeric.StepSize = 1000,
                         string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
