% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromo.Phylome.Custom.R
\name{Chromo.Phylome.Custom}
\alias{Chromo.Phylome.Custom}
\title{Chromo.Phylome.Custom: function to infer a chromosome-specific set of phylogenetic trees}
\usage{
Chromo.Phylome.Custom(
  string.PathParentDir,
  string.PathToFastaFile,
  matrix.WindowCoordinates,
  string.Commands_iqtree
)
}
\arguments{
\item{string.PathParentDir}{Path to parent dir for all analyses}

\item{string.PathToFastaFile}{Path to fasta file}

\item{string.Commands_iqtree}{String used for nucleotide substitution model for IqTree. Leave as "" for model selection}

\item{numeric.WindowSize}{Size of each chromosomal window}

\item{numeric.StepSize}{Spacing of each chromosomal window}
}
\value{
Results Results writted to output files
}
\description{
This function returns a list of the phylogenetic tree models, one for each window set by the user-defined window and step size input parameters
}
\examples{




}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
