% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromo.Phylome.R
\name{Chromo.Phylome}
\alias{Chromo.Phylome}
\title{Chromo.Phylome: function to infer a chromosome-specific set of phylogenetic trees}
\usage{
Chromo.Phylome(
  string.PathParentDir,
  string.PathToFastaFile,
  numeric.WindowSize,
  numeric.StepSize,
  string.Commands_iqtree
)
}
\arguments{
\item{string.PathParentDir}{Path to parent dir for all analyses}

\item{string.PathToFastaFile}{Path to fasta file}

\item{numeric.WindowSize}{Size of each chromosomal window}

\item{numeric.StepSize}{Spacing of each chromosomal window}

\item{string.Commands_iqtree}{String used for nucleotide substitution model for IqTree. Leave as "" for model selection}
}
\value{
Results Results writted to output files
}
\description{
This function returns a list of the phylogenetic tree models, one for each window set by the user-defined window and step size input parameters
}
\examples{




#####################################
# Read example chromosome alignment #
#####################################
String.Path_ExampleChromosomeAlignment <-  system.file("extdata", "Example_Chr10.fasta", package="PhyloWGA")


#########################
# Conduct.ChromoPhylome #
#########################
handle.RESULTS <- Chromo.Phylome(string.PathParentDir = '~/Desktop/',
                                          string.PathToFastaFile = String.Path_ExampleChromosomeAlignment,
                                          numeric.WindowSize = 1000,
                                          numeric.StepSize = 1000,
                                          string.Commands_iqtree = "-m GTR+G")
}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
