% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define.PhyloWGA_Experiment.R
\name{Define.PhyloWGA_Experiment}
\alias{Define.PhyloWGA_Experiment}
\title{Define.PhyloWGA_Experiment: function to generate a matrix of window coordinates for a given alignment}
\usage{
Define.PhyloWGA_Experiment(
  numeric.WindowSize,
  numeric.StepSize,
  numeric.TotalLength
)
}
\arguments{
\item{numeric.WindowSize}{Number of bp for each window, window length}

\item{numeric.StepSize}{Number of bp seperating each window}

\item{numeric.TotalLength}{Total number of bp for the entire alignment}
}
\value{
matrix.WindowCoordinates Matrix of k x 2 dimensions, each row indicating the start and end coordinates for the respective window
}
\description{
This function returns a matrix of the window coordinates for an alignment
}
\examples{


###################################
# Specify experimental parameters #
###################################
matrix.ExperimentalParams <- Define.PhyloWGA_Experiment(numeric.WindowSize = 1000, numeric.StepSize = 500, numeric.TotalLength = 10000)

}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
