% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Execute.PhylogeneticCongruence_concatepillar.R
\name{Execute.PhylogeneticCongruence_concatepillar}
\alias{Execute.PhylogeneticCongruence_concatepillar}
\title{Execute.PhylogeneticCongruence_concatepillar: function to execute concatepillar on a pair of alignments}
\usage{
Execute.PhylogeneticCongruence_concatepillar(
  matrix.WindowAlignment_01,
  matrix.WindowAlignment_02,
  string.PathParentDir,
  numeric.NumberOfCores
)
}
\arguments{
\item{matrix.WindowAlignment_01}{Matrix of alignment for the first window}

\item{matrix.WindowAlignment_02}{Matrix of alignment for the second window}

\item{string.PathParentDir}{Path to parent dir for all analyses}

\item{numeric.NumberOfCores}{Number of cores}
}
\value{
List List of results from concatepillar
}
\description{
This function returns a list of results from concatepillar
}
\examples{


################
# Load depends #
################
library(PhyloWGA)
library(ape)

#####################################
# Read example chromosome alignment #
#####################################
String.Path_ExampleChromosomeAlignment <-  system.file("extdata", "Example_Chr10.fasta", package="PhyloWGA")

########################
# Get alignment length #
########################
numeric.AlignmentLength <- Get.AlignmentLength(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

######################
# Get sequence names #
######################
vector.SequenceNames <- Get.SequenceNames(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

##################################
# Define experimental parameters #
##################################
matrix.ExperimentalParams <- Define.PhyloWGA_Experiment(numeric.WindowSize = 1000, numeric.StepSize = 1000, numeric.TotalLength = numeric.AlignmentLength)

############################
# Extract window alignment #
############################
matrix.WindowAlignment_01 <- Extract.WindowAlignment(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment,
                                                  vector.SequenceNames = vector.SequenceNames,
                                                  vector.Coordinates_ExperimentalLocus = matrix.ExperimentalParams[5,])


####################################################
# Prune taxa that are only missin data from matrix #
####################################################
matrix.Pruned_WindowAlignment_01 <- Prune.MissingDataTaxa(matrix.WindowAlignment = matrix.WindowAlignment_01, numeric.MissingDataThreshold = 0.999)

############################
# Extract window alignment #
############################
matrix.WindowAlignment_02 <- Extract.WindowAlignment(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment,
                                                  vector.SequenceNames = vector.SequenceNames,
                                                  vector.Coordinates_ExperimentalLocus = matrix.ExperimentalParams[6,])


####################################################
# Prune taxa that are only missin data from matrix #
####################################################
matrix.Pruned_WindowAlignment_02 <- Prune.MissingDataTaxa(matrix.WindowAlignment = matrix.WindowAlignment_02, numeric.MissingDataThreshold = 0.999)

##############################
# Estimate model with iqtree #
##############################
Execute.PhylogeneticCongruence_concatepillar(matrix.Pruned_WindowAlignment_01, matrix.Pruned_WindowAlignment_02, string.PathParentDir, numeric.NumberOfCores)

}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
