% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Execute.PhylogeneticInference_iqtree.R
\name{Execute.PhylogeneticInference_iqtree}
\alias{Execute.PhylogeneticInference_iqtree}
\title{Execute.PhylogeneticInference_iqtree: function to estimate a phylogenetic tree model for a window with iqtree}
\usage{
Execute.PhylogeneticInference_iqtree(
  matrix.WindowAlignment,
  string.PathParentDir,
  string.Commands_iqtree
)
}
\arguments{
\item{matrix.WindowAlignment}{Matrix of the alignment for a window}

\item{string.PathParentDir}{Path used for estimating trees with iqtree}
}
\value{
List List containing (1) list of parameter estimates and (2) phylogenetic tree estimate
}
\description{
This function returns a list containing parameter estimates and a phylogenetic tree estimate
}
\examples{




################
# Load depends #
################
library(PhyloWGA)
library(ape)

#####################################
# Read example chromosome alignment #
#####################################
String.Path_ExampleChromosomeAlignment <-  system.file("extdata", "Example_Chr10.fasta", package="PhyloWGA")

########################
# Get alignment length #
########################
numeric.AlignmentLength <- Get.AlignmentLength(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

######################
# Get sequence names #
######################
vector.SequenceNames <- Get.SequenceNames(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

##################################
# Define experimental parameters #
##################################
matrix.ExperimentalParams <- Define.PhyloWGA_Experiment(numeric.WindowSize = 1000, numeric.StepSize = 1000, numeric.TotalLength = numeric.AlignmentLength)

############################
# Extract window alignment #
############################
matrix.WindowAlignment <- Extract.WindowAlignment(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment,
                                                  vector.SequenceNames = vector.SequenceNames,
                                                  vector.Coordinates_ExperimentalLocus = matrix.ExperimentalParams[5,])


####################################################
# Prune taxa that are only missin data from matrix #
####################################################
matrix.Pruned_WindowAlignment <- Prune.MissingDataTaxa(matrix.WindowAlignment = matrix.WindowAlignment, numeric.MissingDataThreshold = 0.999)

##############################
# Estimate model with iqtree #
##############################
Execute.PhylogeneticInference_iqtree(matrix.WindowAlignment = matrix.Pruned_WindowAlignment, string.PathParentDir = '~/Desktop/', string.Commands_iqtree = "-m GTR+G")

}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
