% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.WindowAlignment.R
\name{Extract.WindowAlignment}
\alias{Extract.WindowAlignment}
\title{Extract.WindowAlignment: function to extract an alignment matrix given an input vector of sequences, coordinates, and a path to a fasta file}
\usage{
Extract.WindowAlignment(
  string.PathToFastaFile,
  vector.SequenceNames,
  vector.Coordinates_ExperimentalLocus
)
}
\arguments{
\item{string.PathToFastaFile}{String defining the path to the input fasta file}

\item{vector.SequenceNames}{Vector of sequence names to be extracted}

\item{vector.Coordinates_ExperimentalLocus}{Vector containing two coordinates for the window (start and end)}
}
\value{
matrix.Window_Alignment Matrix containing the alignment for the input window coordinates from the fasta file
}
\description{
This function returns matrix consisting of an extracted alignment
}
\examples{

################
# Load depends #
################
library(PhyloWGA)
library(ape)

#####################################
# Read example chromosome alignment #
#####################################
String.Path_ExampleChromosomeAlignment <-  system.file("extdata", "Example_Chr10.fasta", package="PhyloWGA")

########################
# Get alignment length #
########################
numeric.AlignmentLength <- Get.AlignmentLength(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

######################
# Get sequence names #
######################
vector.SequenceNames <- Get.SequenceNames(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

##################################
# Define experimental parameters #
##################################
matrix.ExperimentalParams <- Define.PhyloWGA_Experiment(numeric.WindowSize = 1000, numeric.StepSize = 1000, numeric.TotalLength = numeric.AlignmentLength)

############################
# Extract window alignment #
############################
matrix.WindowAlignment <- Extract.WindowAlignment(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment,
                                                  vector.SequenceNames = vector.SequenceNames,
                                                  vector.Coordinates_ExperimentalLocus = matrix.ExperimentalParams[5,])


write.dna(x = matrix.WindowAlignment, file = '~/Desktop/ExampleWindow.fasta', colsep = "", format = "fasta")

}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
