% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.AlignmentLength.R
\name{Get.AlignmentLength}
\alias{Get.AlignmentLength}
\title{Get.AlignmentLength: get the total length of an alignment in the input fasta file}
\usage{
Get.AlignmentLength(string.PathToFastaFile)
}
\arguments{
\item{string.PathToFastaFile}{String defining the path to the input fasta file}
}
\value{
numeric.Total_AlignmentLength Number of base pairs in the input fasta file
}
\description{
This function returns the number of bp found in the input fasta file
}
\examples{

#####################################
# Read example chromosome alignment #
#####################################
String.Path_ExampleChromosomeAlignment <-  system.file("extdata", "Example_Chr10.fasta", package="PhyloWGA")

########################
# Get alignment length #
########################
numeric.AlignmentLength <- Get.AlignmentLength(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)
}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
