% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.SequenceNames.R
\name{Get.SequenceNames}
\alias{Get.SequenceNames}
\title{Get.SequenceNames: get a vector of the sequence names provided in a fasta file}
\usage{
Get.SequenceNames(string.PathToFastaFile)
}
\arguments{
\item{string.PathToFastaFile}{String defining the path to the input fasta file}
}
\value{
vector.SequenceNames Vector containing the set of sequence names obtained from fasta file
}
\description{
This function returns a vector of sequence names
}
\examples{



#####################################
# Read example chromosome alignment #
#####################################
String.Path_ExampleChromosomeAlignment <-  system.file("extdata", "Example_Chr10.fasta", package="PhyloWGA")

########################
# Get alignment length #
########################
vector.SequenceNames <- Get.SequenceNames(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)
}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
