% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organize.ParallelPhyloWGA.R
\name{Organize.ParallelPhyloWGA}
\alias{Organize.ParallelPhyloWGA}
\title{Organize.ParallelPhyloWGA: function to partition and organize subdirectories for parallel WGA analyses}
\usage{
Organize.ParallelPhyloWGA(
  numeric.NumberSubsets,
  string.PathToFastaFile,
  string.PathParentDir,
  string.Analysis,
  numeric.WindowSize,
  string.Commands_iqtree,
  numeric.StepSize,
  numeric.NumberOfCores
)
}
\arguments{
\item{numeric.NumberSubsets}{Number of subsets for the original alignment and analyses}

\item{string.PathToFastaFile}{String defining the path to the input fasta file}

\item{string.PathParentDir}{Path to parent dir for all analyses}

\item{numeric.WindowSize}{Size of each chromosomal window}

\item{numeric.StepSize}{Spacing of each chromosomal window}

\item{numeric.NumberOfCores}{Number of cores}
}
\value{
XXX XXX
}
\description{
This function returns a XXX
}
\examples{


}
\keyword{coalescent}
\keyword{genome,}
\keyword{parallel}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
