% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prune.MissingDataTaxa.R
\name{Prune.MissingDataTaxa}
\alias{Prune.MissingDataTaxa}
\title{Prune.MissingDataTaxa: function to pruned matrix (high-missing data sequences removed)}
\usage{
Prune.MissingDataTaxa(matrix.WindowAlignment, numeric.MissingDataThreshold)
}
\arguments{
\item{matrix.WindowAlignment}{Matrix of input chromosomal window alignment}

\item{numeric.MissingDataThreshold}{Numeric (proportion) of missing data permited of any one sequence}
}
\value{
matrix.Pruned_WindowAlignment Matrix of window alignment with sequences with % missing data above the threshold removed
}
\description{
This function returns a matrix that has been pruned of sequences with missing data above a given threshold
}
\examples{

################
# Load depends #
################
library(PhyloWGA)
library(ape)

#####################################
# Read example chromosome alignment #
#####################################
String.Path_ExampleChromosomeAlignment <-  system.file("extdata", "Example_Chr10.fasta", package="PhyloWGA")

########################
# Get alignment length #
########################
numeric.AlignmentLength <- Get.AlignmentLength(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

######################
# Get sequence names #
######################
vector.SequenceNames <- Get.SequenceNames(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

##################################
# Define experimental parameters #
##################################
matrix.ExperimentalParams <- Define.PhyloWGA_Experiment(numeric.WindowSize = 1000, numeric.StepSize = 1000, numeric.TotalLength = numeric.AlignmentLength)

############################
# Extract window alignment #
############################
matrix.WindowAlignment <- Extract.WindowAlignment(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment,
                                                  vector.SequenceNames = vector.SequenceNames,
                                                  vector.Coordinates_ExperimentalLocus = matrix.ExperimentalParams[5,])


####################################################
# Prune taxa that are only missin data from matrix #
####################################################
matrix.Pruned_WindowAlignment <- Prune.MissingDataTaxa(matrix.WindowAlignment = matrix.WindowAlignment, numeric.MissingDataThreshold = 0.999)

}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
