% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.OutputFile_iqtree.R
\name{Read.OutputFile_iqtree}
\alias{Read.OutputFile_iqtree}
\title{Read.OutputFile_iqtree: function to extract a list of parameter estimates from an iqtree outputfile}
\usage{
Read.OutputFile_iqtree(string.Path_OutputFile_iqtree)
}
\arguments{
\item{string.Path_OutputFile_iqtree}{String defining path to the output iqtree file}
}
\value{
List List of containing estimates of the GTR+G model
}
\description{
This function returns a list of the parameter estimates
}
\examples{


################
# Load depends #
################
library(PhyloWGA)
library(ape)

#####################################
# Read example chromosome alignment #
#####################################
String.Path_ExampleChromosomeAlignment <-  system.file("extdata", "Example_Chr10.fasta", package="PhyloWGA")

########################
# Get alignment length #
########################
numeric.AlignmentLength <- Get.AlignmentLength(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

######################
# Get sequence names #
######################
vector.SequenceNames <- Get.SequenceNames(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment)

##################################
# Define experimental parameters #
##################################
matrix.ExperimentalParams <- Define.PhyloWGA_Experiment(numeric.WindowSize = 1000, numeric.StepSize = 1000, numeric.TotalLength = numeric.AlignmentLength)

############################
# Extract window alignment #
############################
matrix.WindowAlignment <- Extract.WindowAlignment(string.PathToFastaFile = String.Path_ExampleChromosomeAlignment,
                                                  vector.SequenceNames = vector.SequenceNames,
                                                  vector.Coordinates_ExperimentalLocus = matrix.ExperimentalParams[5,])


####################################################
# Prune taxa that are only missin data from matrix #
####################################################
matrix.Pruned_WindowAlignment <- Prune.MissingDataTaxa(matrix.WindowAlignment = matrix.WindowAlignment, numeric.MissingDataThreshold = 0.999)

##############################
# Estimate model with iqtree #
##############################
Execute.PhylogeneticInference_iqtree(matrix.WindowAlignment = matrix.Pruned_WindowAlignment, string.PathParentDir = '~/Desktop/')

####################
# Read output file #
####################
Read.OutputFile_iqtree(string.Path_OutputFile_iqtree = '~/Desktop/PhylogeneticInference_iqtree/WindowAlignment.fa.iqtree')

}
\keyword{coalescent}
\keyword{genome,}
\keyword{phylogenetics,}
\keyword{species}
\keyword{theory,}
\keyword{tree,}
\keyword{whole}
